fix bug

+++ src/metric.py
@@ -17,7 +17,7 @@ class F1:
         target = torch.cat(self.targets)
         scores = []
         for threshold in np.linspace(0, 1, num=100, endpoint=False):
-            predict = origin_predict > threshold
+            predict = (origin_predict > threshold).float()
             tp = (predict*target).sum(dim=0)  #shape (28,)
             precision = tp/(predict.sum(dim=0) + 1e-8)
             recall = tp/(target.sum(dim=0) + 1e-8)
